adding tasks works but with hacky implementation
