+++ src/add_task/add_task.jsx
@@ -5,8 +5,8 @@ export default function AddTask(props) {
 
     async function handleSubmit(event) {
         event.preventDefault()
-        let result = await props.connector.addTask(taskInput.current.value)
-        console.log(result)
+        await props.connector.addTask(taskInput.current.value)
+        props.getData();
     }
     
     return (
